Move clean_text to global utils function
